requirements.txt
crewai[tools]>=0.30.0
pandas>=2.0.0
streamlit>=1.30.0
plotly>=5.18.0
python-dotenv>=1.0.0
pydantic>=2.5.0
loguru>=0.7.0
openpyxl>=3.1.0
scikit-learn>=1.3.0
ydata-profiling>=4.6.0
Pillow>=10.0.0
statsmodels>=0.14.0
